Locking down fields and mutations

+++ src/resolvers/Mutation.js
@@ -59,8 +59,14 @@ const Mutation = {
       }
     });
 
-    if(!postExists) {
+    if (!postExists) {
       throw new Error('Unable to update post');
+    }
+
+    const isPublished = await prisma.exists.Post({id: id, published: true});
+
+    if (!isPublished && published === false) {
+      await prisma.mutation.deleteManyComments({post: id});
     }
 
     return prisma.mutation.updatePost({
@@ -93,9 +99,18 @@ const Mutation = {
       }
     }, info);
   },
-  createComment(parent, args, {prisma, request }, info) {
+  async createComment(parent, args, {prisma, request }, info) {
     const userId = getUserId(request);
     const {text, post} = args.data;
+    const postExists = await prisma.exists.Post({
+      id: post,
+      published: true
+    });
+
+    if (!postExists) {
+      throw new Error("Unable to find post");
+    }
+
     return prisma.mutation.createComment({
       data: {
         text,
@@ -112,16 +127,40 @@ const Mutation = {
     }
     }, info)
   },
-  deleteComment(parent, args, {prisma}, info) {
+  async deleteComment(parent, args, {prisma, request}, info) {
+    const userId = getUserId(request);
     const {id} = args;
+    const commentExists = await prisma.exists.Comment({
+      id,
+      author: {
+        id: userId
+      }
+    });
+
+    if(!commentExists) {
+      throw new Error('Unable to delete comment');
+    }
+
     return prisma.mutation.deleteComment({
       where: {
         id
       }
     }, info);
   },
-  updateComment(parent, args, {prisma}, info) {
+  async updateComment(parent, args, {prisma, request}, info) {
+    const userId = getUserId(request);
     const {id, data} = args;
+    const commentExists = await prisma.exists.Comment({
+      id,
+      author: {
+        id: userId
+      }
+    });
+
+    if(!commentExists) {
+      throw new Error('Unable to update comment');
+    }
+
     return prisma.mutation.updateComment({
       where: {
         id
